ThemeQueryManager: Move isThemeMatchingQuery() to state/themes/util
It's no longer needed by ThemeQueryManager.
